\name{Val_Estadisticos}
\alias{Val_Estadisticos}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Estadigrafos de un conjunto de observaciones
}
\description{
Esta función permite calcular los estadígrafos de un DataFrame
}
\usage{
Val_Estadisticos(datos)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datos}{
Dataframe con las observaciones
}
}
\details{

}
\value{
\item{Tabla}{Tabla con los estadígrafos de todas las variables dentro del DataFrame}
}
\references{
Consultar las diapositivas CG2 y las notas del curso
}
%\author{
%%  ~~who you are~~
%}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
%\examples{
%##---- Should be DIRECTLY executable !! ----
%##-- ==>  Define data, use random,
%##--	or do  help(data=index)  for the standard data sets.
%
%## The function is currently defined as
%function (x)
%{
%  }
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
